+++ ntex-io/src/dispatcher.rs
@@ -250,7 +250,7 @@ where
                     // call service
                     let shared = slf.shared.clone();
                     shared.inflight.set(shared.inflight.get() + 1);
-                    let fut = shared.service.static_call(item);
+                    let fut = shared.service.call(item).into_static();
                     spawn(async move {
                         let result = fut.await;
                         shared.handle_result(result, &shared.io);
@@ -276,7 +276,7 @@ where
                     // call service
                     let shared = slf.shared.clone();
                     shared.inflight.set(shared.inflight.get() + 1);
-                    let fut = shared.service.static_call(item);
+                    let fut = shared.service.call(item).into_static();
                     spawn(async move {
                         let result = fut.await;
                         shared.handle_result(result, &shared.io);