+++ ntex-service/CHANGES.md
@@ -1,8 +1,10 @@
 # Changes
 
-## [1.2.0-beta.3] - 2023-06-20
+## [1.2.0-beta.3] - 2023-06-21
 
-* Add Container::static_call(), returns future that is suitable for spawning into runtime.
+* Add custom ContainerCall future
+
+* Allow to turn ContainerCall to static
 
 ## [1.2.0-beta.2] - 2023-06-19